### 2020-06-18 #2356 @whedon
[PRE REVIEW]: textnets: A Python package for text analysis with networks
**Submitting author:** @jboynyc (<a href="http://orcid.org/0000-0003-2118-4702">John Boy</a>)
**Repository:** <a href="https://github.com/jboynyc/textnets" target ="_blank">https://github.com/jboynyc/textnets</a>
**Version:** v.0.4.2
