### 2020-06-18 086ebe2
change how nodes are scaled to work for differently skewed distributions
