### 2020-06-18 c316667
fix: typo in documentation
