### 2020-06-18 #2356 @jboynyc
Sounds good, thanks! Possible reviewers include contributors to [cbail/textnets](https://github.com/cbail/textnets), alexhanna, cmaimone, tresoldi, and nniiicc. I'll update this comment if I find any more.
