### 2020-06-18 aaecb01
complete change of centrality measures from lists into series
