### 2020-06-19 6d4b63c
use language short codes in tutorial
