### 2020-06-19 d2d9839
update installation and contribution docs
