+++ docs/installation.rst
@@ -18,8 +18,7 @@ In a `virtual environment`_, run::
 
    $ pip install textnets
 
-.. _`virtual environment`:
-https://packaging.python.org/tutorials/installing-packages/#creating-virtual-environments
+.. _`virtual environment`: https://packaging.python.org/tutorials/installing-packages/#creating-virtual-environments
 
 This is the preferred method to install **textnets**, as it always installs
 the most recent stable release.
@@ -27,8 +26,7 @@ the most recent stable release.
 If you don't have pip installed, the `Python installation guide`_ can guide you
 through the process.
 
-.. _Python installation guide:
-http://docs.python-guide.org/en/latest/starting/installation/
+.. _Python installation guide: http://docs.python-guide.org/en/latest/starting/installation/
 
 Most likely you also have to install an appropriate `language model`_ by
 issuing a command like::
