### 2020-06-19 8a433a9
fix: tutorial typos
