### 2020-06-19 9a1d39e
reorganize options in plot method
