### 2020-06-19 04a41e9
change iteration parameter: run until there is no improvement
