### 2020-06-19 8d4139a
improve docstring in tests
