### 2020-06-25 b1b6087
add option to normalize (lemmatize) noun chunks
