### 2020-06-25 7612abb
fix venv documentation in contributor docs
