### 2020-06-27 b3d76de
add option to filter node and edge labels; closes #12
