### 2020-06-29 d1f32e1
add docstrings documenting label filter options
