### 2020-06-29 9552f1c
add useful string representation of `Corpus` object
