### 2020-06-29 d2e20ad
fix typos in documentation
