+++ HISTORY.rst
@@ -2,6 +2,15 @@
 History
 =======
 
+0.4.5 (2020-06-29)
+------------------
+* `Textnet.plot` and `TextnetBase.plot` now accept arguments to selectively
+  suppress node or edge labels. ``node_label_filter`` and ``edge_label_filter``
+  take a function that is mapped to the iterator of nodes and edges. Only nodes
+  or edges for which the function returns ``True`` are displayed in the plot.
+* `Corpus` now has a useful string representation.
+* Documentation updates, particularly to show the label filter functionality.
+
 0.4.4 (2020-06-19)
 ------------------
 
