### 2020-06-29 c407bae
start to restructure unit tests; addresses #11
