+++ tests/conftest.py
@@ -1,0 +1,17 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+"""Configure tests for `textnets` package."""
+
+import pytest
+from textnets import Corpus, Textnet, examples
+
+
+@pytest.fixture(scope="session")
+def testdata():
+    return examples.moon_landing
+
+
+@pytest.fixture(scope="session")
+def corpus(testdata):
+    return Corpus(testdata)
