### 2020-07-01 286310b
update spacy dependency and list of language models
