### 2020-07-01 ef41142
fix how nodes are scaled
