### 2020-07-01 a13bf33
update changelog, update language model used for testing
