+++ tox.ini
@@ -15,7 +15,7 @@ commands = flake8 textnets tests
 setenv =
     PYTHONPATH = {toxinidir}
 deps =
-    https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz#egg=en_core_web_sm-2.2.5
+    https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.0/en_core_web_sm-2.3.0.tar.gz#egg=en_core_web_sm-2.3.0
     -e .[test]
 ; If you want to make tox run the tests with the same versions, create a
 ; requirements.txt with the pinned versions and uncomment the following line:
