### 2020-07-01 864901c
update changelog, update language model used for testing
