### 2020-07-09 4823792
handle series with missing data (closes #13)
