### 2020-07-09 3ea4636
fix: remove walrus to ensure compatibility with Python 3.7
