### 2020-07-09 6096e66
improve documentation of dependencies
