### 2020-07-09 3bde75d
refactoring warnings
