### 2020-07-09 3d23a3f
add test for a Czech language corpus using blank model
