### 2020-07-12 #2356 @cmaimone
I've got 2 open reviews. I'd like to complete those before committing to
any additional ones.

On Sun, Jul 12, 2020, 12:05 PM George K. Thiruvathukal <
notifications@github.com> wrote:

> @alexhanna <https://github.com/alexhanna> and @cmaimone
> <https://github.com/cmaimone>, I am in need of 2 reviewers for this JOSS
> submission. Are either or both of you willing to serve as reviewers?
>
