+++ paper.md
@@ -23,7 +23,7 @@ Owing to the ever-expanding digitization of social life, social scientists need
 computational tools to make sense of vasts amounts of unstructured data.
 Electronic text, in particular, is a growing area of interest thanks to the
 social and cultural insights lurking in social media posts, digitized corpora,
-and web content, among other troves [@Evans2016].
+and web content, among other troves [@Evans2016; @Ignatow2015].
 
 This package aims to fill that need. `textnets` represents collections of texts
 as networks of documents and words, which provides novel and exciting
