+++ paper.bib
@@ -87,3 +87,13 @@
   pages = {eaaq1360},
   doi = {10.1126/sciadv.aaq1360},
 }
+@article{Ignatow2015,
+  author = {Gabe Ignatow},
+  title = {Theoretical Foundations for Digital Text Analysis},
+  journal = {Journal for the Theory of Social Behaviour},
+  volume = {46},
+  number = {1},
+  year = {2015},
+  pages = {104--120},
+  doi = {10.1111/jtsb.12086},
+}
