### 2020-07-15 7a3aa89
fix: the bipartite network is not a projection
