### 2020-07-15 c4a0496
clarify language support in tutorial
