### 2020-07-15 693d0fd
clean up tutorial text
