### 2020-07-15 84b376f
add `color_clusters` option to plot methods
