+++ textnets/network.py
@@ -23,6 +23,7 @@ import igraph as ig
 import leidenalg as la
 from scipy.integrate import quad
 
+from .viz import TextnetPalette
 from .fca import FormalContext
 
 
@@ -215,6 +216,7 @@ class TextnetBase:
     def _plot(
         self,
         show_clusters: Union[bool, ig.clustering.VertexClustering] = False,
+        color_clusters: Union[bool, ig.clustering.VertexClustering] = False,
         label_edges: bool = False,
         scale_nodes_by: Optional[str] = None,
         node_label_filter: Optional[Callable[[ig.Vertex], bool]] = None,
@@ -238,12 +240,30 @@ class TextnetBase:
                 kwargs.setdefault("mark_groups", show_clusters)
             else:
                 kwargs.setdefault("mark_groups", self.clusters)
+        if color_clusters:
+            if isinstance(color_clusters, ig.clustering.VertexClustering):
+                kwargs.setdefault(
+                    "vertex_color",
+                    [
+                        TextnetPalette(color_clusters.n)[c]
+                        for c in color_clusters.membership
+                    ],
+                )
+            else:
+                kwargs.setdefault(
+                    "vertex_color",
+                    [
+                        TextnetPalette(self.clusters.n)[c]
+                        for c in self.clusters.membership
+                    ],
+                )
+        else:
+            kwargs.setdefault(
+                "vertex_color",
+                ["orangered" if v else "dodgerblue" for v in self.node_types],
+            )
         kwargs.setdefault(
             "vertex_shape", ["circle" if v else "square" for v in self.node_types]
-        )
-        kwargs.setdefault(
-            "vertex_color",
-            ["orangered" if v else "dodgerblue" for v in self.node_types],
         )
         kwargs.setdefault(
             "vertex_frame_color", ["black" if v else "white" for v in self.node_types]
@@ -393,7 +413,11 @@ class Textnet(TextnetBase, FormalContext):
 
         Parameters
         ----------
-        show_clusters : bool of VertexClustering, optional
+        color_clusters : bool or VertexClustering, optional
+            Color nodes according to clusters detected by Leiden algorithm
+            (default: False). Alternately a clustering object generated by
+            another community detection algorithm can be passed.
+        show_clusters : bool or VertexClustering, optional
             Mark clusters detected by Leiden algorithm (default: False).
             Alternately a clustering object generated by another community
             detection algorithm can be passed.
@@ -494,7 +518,11 @@ class ProjectedTextnet(TextnetBase):
 
         Parameters
         ----------
-        show_clusters : bool, optional
+        color_clusters : bool or VertexClustering, optional
+            Color nodes according to clusters detected by Leiden algorithm
+            (default: False). Alternately a clustering object generated by
+            another community detection algorithm can be passed.
+        show_clusters : bool or VertexClustering, optional
             Mark clusters detected by Leiden algorithm (default: False).
             Alternately a clustering object generated by another community
             detection algorithm can be passed.
