### 2020-07-15 db45f51
remove outdated comments
