### 2020-07-15 43217e3
show off color_clusters in tutorial
