### 2020-07-16 48e2721
revise tutorial to show off all plotting features
