### 2020-07-19 #2356 @sara-02
Hey @jboynyc @gkthiruvathukal I am looking for more packages to review, and even though I have not contributed to textnets, I have knowledge about spacy and text analysis and community detection algorithms.
I would be happy to take this up.
