### 2020-09-06 jboynyc/textnets#14 @tresoldi
Add `cairocffi` as a dependency
* Textnets version: 0.4.9
* Python version: 3.7
* Operating System: Linux

### Description

In most environments created afresh, no engine for rendering the Cairo output will be available, which means the the examples from the tutorial will not run out-of-the-box.

Also considering how Cairo can be difficult to install in some cases, I suggest adding `cairocffi` as a dependency or, at least, clearly stating the alternatives.

### What I Did

```
pip install cairocffi
```
