### 2020-09-06 #2594 @tresoldi
@jboynyc great library, thank you! I can see myself using it for at least one project I have in mind.

I've just filed an issue on the only real problem I saw. Some additional points/comments here:

- Given your audience, it might be valid to offer the library as a conda package as well
- I suggest putting
