### 2020-09-07 jboynyc/textnets#15 @jboynyc
Create conda package
A conda package might help some researchers get started with textnets, as [recommended](https://github.com/openjournals/joss-reviews/issues/2594#issuecomment-687823038) by @tresoldi.

Possibly helpful for deploying a conda package using Travis: [stackoverflow question](https://stackoverflow.com/questions/63424864/deploying-python-package-on-anaconda-with-travisci), [example config](https://gist.github.com/yoavram/05a3c04ddcf317a517d5).
