### 2020-09-08 #2594 @jboynyc
Thanks, @tresoldi! I will push a new version of the paper by the end of the week adding a more detailed statement of need to the background section and doing my best to fix figure placement. I'll let you know when I'm ready.

I also really like your idea of creating a Conda package. I've started looking into it.
