+++ paper.bib
@@ -97,3 +97,14 @@
   pages = {104--120},
   doi = {10.1111/jtsb.12086},
 }
+@article{Benoit2018,
+  author = {Kenneth Benoit and Kohei Watanabe and Haiyan Wang and Paul Nulty and Adam Obeng and Stefan M{\"u}ller and Akitaka Matsuo},
+  title = {quanteda},
+  subtitle = {An {R} package for the quantitative analysis of textual data},
+  journal = {Journal of Open Source Software},
+  volume = {3},
+  number = {30},
+  year = {2018},
+  pages = {774},
+  doi = {10.21105/joss.00774},
+}
