+++ paper.md
@@ -13,7 +13,7 @@ authors:
 affiliations:
  - name: Assistant Professor, Faculty of Social Sciences, Leiden University
    index: 1
-date: 16 June 2020
+date: 14 September 2020
 bibliography: paper.bib
 ---
 
