### 2020-09-14 bcfa927
rewrite opening paragraphs
