### 2020-09-14 da3c95e
add citation to paper caption
