+++ paper.bib
@@ -108,3 +108,9 @@
   pages = {774},
   doi = {10.21105/joss.00774},
 }
+@online{Boy2020,
+  author = {John D. Boy},
+  title = {Enemies, Foreign and Partisan},
+  url = {https://www.jboy.space/blog/enemies-foreign-and-partisan.html},
+  date = {2020-06-11},
+}
