+++ paper.bib
@@ -71,7 +71,7 @@
   booksubtitle = {Players, Agents and Agendas},
   booktitleaddon = {Proceedings of the 20th International Conference on Electronic Publishing},
   editor = {Fernando Loizides and Birgit Schmidt},
-  location = {G{\"o}ttingen},
+  address = {Amsterdam},
   publisher = {IOS Press},
   year = {2016},
   pages = {87--90},
