### 2020-09-14 3b16259
move figure to improve placement
