### 2020-09-14 b9fed53
move figure for better placement
