### 2020-09-14 #2594 @tresoldi
All my points were addressed, only issue https://github.com/jboynyc/textnets/issues/14 still pending.
