### 2020-09-14 8e07647
update docs on pycairo dependency (#14)
