### 2020-09-14 5d491be
pycairo no longer needed as separate test dependency
