### 2020-09-14 jboynyc/textnets#14 @jboynyc
`pycairo` is now included as a dependency to make use of plotting features easier (0e8af66). I updated the installation docs to reflect this, and to direct users for whom the pycairo dependency does not work to install `cairoffi` using pip or conda (8e07647). I'd appreciate your feedback on whether that is clear/helpful, @tresoldi.

I will merge this into trunk and create a new release soon.

I agree that creating a Conda package (#15) would probably still be the easiest way to get Windows users on board.

**Edit:** Apologies, I just understood that you suggested requiring *a drop-in replacement* for pycairo. I will test that option now.
