### 2020-09-14 5adba03
switch plotting dependency from pycairo to cairocffi
