### 2020-09-14 4898b4d
bump leidenalg dependency version
