### 2020-09-14 jboynyc/textnets#14 @tresoldi
This is great, thanks! As for the suggestion, it is mostly in order to follow the journal instructions that users should be able to follow the docs to get the results that are demonstrated. A drop-in replacement for pycairo can be taken as a suggestion, but not as a request, I'm happy with these instructions.

As for Conda, it is also a suggestion, but approving the manuscript absolutely does not depend on that. In fact, I suppose all is done, once you release and make a note in the JOSS issue I'm happy to give my ok. :)
