### 2020-09-14 ddc5b0f
rewrite installation docs, this time mentioning CFFI
