### 2020-09-14 ee52f8e
update pandas docs link
