### 2020-09-14 jboynyc/textnets#14 @jboynyc
Thanks! I did make the change to cairocffi because it looks less painful as a cross-platform dependency (5adba03a3831fc74cf58db372a26d370a925cbcc, ddc5b0fdb175203ab927b540cbb6c465727e1166).
