### 2020-09-14 #2594 @jboynyc
Release [0.4.10](https://github.com/jboynyc/textnets/releases/tag/v0.4.10) closes jboynyc/textnets#14.

Many thanks @tresoldi for the helpful and encouraging review!
