### 2020-09-14 10a58eb
update leidenalg reference to use DOI for all versions
