+++ paper.bib
@@ -31,10 +31,8 @@
 @software{Traag2020,
   author = {Vincent Traag},
   title = {leidenalg},
-  month = apr,
   year = 2020,
-  version = {0.8.0},
-  doi = {10.5281/zenodo.3779234},
+  doi = {10.5281/zenodo.1469356},
 }
 @article{Csardi2006,
   author = {G{\'a}bor Cs{\'a}rdi and Tam{\'a}s Nepusz},
