### 2020-09-14 jboynyc/textnets#15 @jboynyc
conda-forge recipe under review at https://github.com/conda-forge/staged-recipes/pull/12656
