### 2020-09-14 jboynyc/textnets#15 @jboynyc
now [included](https://anaconda.org/conda-forge/textnets) in conda-forge!
