### 2020-09-14 830eb4f
update install documentation with conda-forge instructions
