+++ docs/installation.rst
@@ -4,15 +4,55 @@
 Installation
 ============
 
-**textnets** is in the `Python Package Index`_, so it can be installed using
-`pip`_.
+**textnets** is included in `conda-forge`_ and the `Python Package Index`_, so
+it can either be installed using `conda`_ or `pip`_.
 
+.. _`conda-forge`: https://anaconda.org/conda-forge/textnets/
 .. _`Python Package Index`: https://pypi.org/project/textnets/
+.. _conda: https://conda.io/
 .. _pip: https://pip.pypa.io
 
 .. note::
 
    Please note that **textnets** requires Python 3.7 or newer to run.
+
+Using conda
+-----------
+
+This is the preferred method for most users. The `Anaconda Python
+distribution`_ is an easy way to get up and running with Python, especially if
+you are on a Mac or Windows system.
+
+.. _Anaconda Python distribution: https://www.anaconda.com/products/individual
+
+Once it is installed you can use its package manager ``conda`` to install
+**textnets**::
+
+   $ conda install -c conda-forge textnets
+
+This tells conda to install **textnets** from the conda-forge channel.
+
+If you don't know how to enter this command, you can use the Anaconda Navigator
+instead. It provides a graphical interface that allows you to install new
+packages.
+
+1. Go to the **Environments** tab.
+2. Click the **Channels** button.
+3. Click the **Add** button.
+4. Enter the channel url https://conda.anaconda.org/conda-forge/
+5. Hit your keyboard's Enter key.
+6. Click the **Update channels** button.
+7. Now you can install **textnets** in a new environment. (Make sure the
+   package filter on the **Environments** tab is set to "all.")
+
+Using pip
+---------
+
+Alternately, if you already have Python installed, you can use its package
+manger. (If you don't have pip installed yet, the `Python installation guide`_
+can guide you through the process.)
+
+.. _Python installation guide: http://docs.python-guide.org/en/latest/starting/installation/
 
 In a `virtual environment`_, run::
 
@@ -20,20 +60,11 @@ In a `virtual environment`_, run::
 
 .. _`virtual environment`: https://packaging.python.org/tutorials/installing-packages/#creating-virtual-environments
 
-This is the preferred method to install **textnets**, as it always installs
-the most recent stable release.
-
-If you don't have pip installed, the `Python installation guide`_ can guide you
-through the process.
-
-.. _Python installation guide: http://docs.python-guide.org/en/latest/starting/installation/
-.. _conda: https://conda.io/
-
 Plotting
 --------
 
 **textnets** installs the `Cairo`_ graphics library as a dependency for
-plotting using the `cairocffi`_ package. In some cases you may have to `install
+plotting using the `cairocffi`_ package. In rare cases you may have to `install
 CFFI`_ separately for plotting to work.
 
 .. _Cairo: https://www.cairographics.org/
