### 2020-09-14 a16fbde
update readme with conda-forge info
