### 2020-09-14 bfc99f0
add change log entry for documentation update
