### 2020-09-16 9459926
mention conda-forge installation option
