### 2020-09-16 #2594 @jboynyc
I made a minor update to the paper noting the ability to [install from conda-forge](https://anaconda.org/conda-forge/textnets).
