### 2020-10-14 #2594 @sara-02
@jboynyc I have a query regarding this point
```
Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
```
I could not find a command for the same, have added an issue here: https://github.com/jboynyc/textnets

Apart from this point, the rest of the checklist is complete from my side as well.
