### 2020-10-14 #2594 @jboynyc
- Latest tagged release is [0.4.10](https://github.com/jboynyc/textnets/releases/tag/v0.4.10)
- Zenodo DOI for that release is [10.5281/zenodo.4028610](https://doi.org/10.5281/zenodo.4028610)
