### 2020-10-15 #2594 @danielskatz
@gkthiruvathukal - when the paper is a in branch, you need to accept it from the branch:
