### 2020-10-15 #2594 @danielskatz
👋 @sara-02 - above you say that the unchecked item in your list is satisfied, so I am going to move the acceptance forward, but if you could check off the item itself, it will be slightly cleaner.
