### 2020-10-15 5943ea5
minor proofreading fixes to paper
