+++ paper.bib
@@ -62,12 +62,12 @@
   doi = {10.1038/s41592-019-0686-2},
 }
 @inproceedings{Kluyver2016,
-  author = {Thomas Kluyver Thomas and Benjamin Ragan-Kelley and Fernando P{\'e}rez and Brian Granger and Matthias Bussonnier and Jonathan Frederic and Kyle Kelley and Jessica Hamrick and Jason Grout and Sylvain Corlay and Paul Ivanov and Dami{\'a}n Avila and Safia Abdalla and Carol Willing and Jupyter Development Team},
+  author = {Thomas Kluyver and Benjamin Ragan-Kelley and Fernando P{\'e}rez and Brian Granger and Matthias Bussonnier and Jonathan Frederic and Kyle Kelley and Jessica Hamrick and Jason Grout and Sylvain Corlay and Paul Ivanov and Dami{\'a}n Avila and Safia Abdalla and Carol Willing and {Jupyter Development Team}},
   title = {Jupyter Notebooks},
   subtitle = {A publishing format for reproducible computational workflows},
   booktitle = {Positioning and Power in Academic Publishing},
   booksubtitle = {Players, Agents and Agendas},
-  booktitleaddon = {Proceedings of the 20th International Conference on Electronic Publishing},
+  booktitleaddon = {{Proceedings of the 20th International Conference on Electronic Publishing}},
   editor = {Fernando Loizides and Birgit Schmidt},
   address = {Amsterdam},
   publisher = {IOS Press},
