### 2020-10-15 #2594 @danielskatz
👋 @jboynyc - Some fixes are needed in the paper: https://github.com/jboynyc/textnets/pull/19 and https://github.com/jboynyc/textnets/pull/20
