### 2020-10-15 #2594 @jboynyc
> wave @jboynyc - Some fixes are needed in the paper: [jboynyc/textnets#19](https://github.com/jboynyc/textnets/pull/19) and [jboynyc/textnets#20](https://github.com/jboynyc/textnets/pull/20)

Merged! Thanks for catching those.
